codelore-backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
gitpython==3.1.40
pydriller==2.5.1
openai
requests
python-dotenv 